+++ crates/api_client/src/models/media_content_moderation_rejected_reason_category.rs
@@ -14,11 +14,11 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct MediaContentModerationRejectedReasonCategory {
     #[serde(rename = "value")]
-    pub value: i64,
+    pub value: i32,
 }
 
 impl MediaContentModerationRejectedReasonCategory {
-    pub fn new(value: i64) -> MediaContentModerationRejectedReasonCategory {
+    pub fn new(value: i32) -> MediaContentModerationRejectedReasonCategory {
         MediaContentModerationRejectedReasonCategory {
             value,
         }
