+++ crates/api_client/src/models/profile_sync_version.rs
@@ -14,11 +14,11 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct ProfileSyncVersion {
     #[serde(rename = "version")]
-    pub version: i64,
+    pub version: i32,
 }
 
 impl ProfileSyncVersion {
-    pub fn new(version: i64) -> ProfileSyncVersion {
+    pub fn new(version: i32) -> ProfileSyncVersion {
         ProfileSyncVersion {
             version,
         }
